requirements.txt
streamlit
radon
openai
python-dotenv
reportlab
matplotlib
